requirements.txt
agentops
anthropic
python-dotenv
